Resolve the depreciation issue of the --proxy parameter
The last script worked great but with the newer version of Keycloak, it says that it will remove the --proxy feature in the future. After a lot of research, I found ways to resolve this issue and get a very similar functionality after modifying the key parameters of the start command.