Resolved issue of database not being utilized
I recently realized that Keycloak wasn't using the supplied database. After doing quite a research I understood that you have to initialize the database by splitting the connection parameters. For this purpose, I created a new file named config.sh in the previous commit which would convert the whole database URL string into small bits and parts that Keycloak is able to utilize to connect to the database.